backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = []
requires-python = ">=3.12,<4.0"
dependencies = [
    "beanie (>=2.0.0,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
